+++ Source/Engine/simulation.cpp
@@ -141,7 +141,8 @@ bool Simulation::Run( void ) {
 
 		// Update the fps once per second
 		if( (Timer::GetTicks() - fpsTS) >1000 ) { 
-			Simulation::currentFPS = 1000.0 * ((float)fpsCount / (Timer::GetTicks() - fpsTS));
+			Simulation::currentFPS = static_cast<float>(1000.0 *
+					((float)fpsCount / (Timer::GetTicks() - fpsTS)));
 			fpsTS = Timer::GetTicks();
 			fpsCount = 0;
 		}
