+++ Source/Utilities/coordinate.cpp
@@ -100,11 +100,12 @@ float Coordinate::GetAngle() {
 	// 0 is right, 90 is up
 	// Due to the way coordinates are displayed, use negative Y
 	Trig *trig = Trig::Instance();
-	return trig->RadToDeg( atan2( -m_y , m_x ) );
+	return static_cast<float>(trig->RadToDeg( static_cast<float>(atan2( static_cast<float>(-m_y)
+					,static_cast<float>( m_x ))) ));
 }
 
 float Coordinate::GetMagnitude() {
-	return sqrt( m_y*m_y + m_x*m_x );
+	return static_cast<float>(sqrt( m_y*m_y + m_x*m_x ));
 }
 
 ostream& operator<<(ostream &out, const Coordinate &c) {
